Provides a docker implementation of a AA-SI test
This PR introduces the initial steps toward implementing Docker for Hannan's library and tests. This is not in response to any specific feature request, but rather a proactive effort to integrate with the Docker containers provided via our workstations. The primary goal is to enhance plug-and-play functionality and automation.

As part of the sprint, I would like to acknowledge that Hannan’s library has been working effectively on my end. The Docker implementation serves as confirmation that it is functioning correctly, and the automation component is included to demonstrate that the process can be automated for greater efficiency.

Lastly, this Dockerfile is intended to replace the images provided through our workstations, as those require significant configuration. This implementation aims to simplify and streamline that process.